% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.utils.R
\name{extract.metadata.asd}
\alias{extract.metadata.asd}
\title{Extract metadata from raw binary ASD files. Called from extract.metadata}
\usage{
extract.metadata.asd(file.dir, out.dir, spec.file.ext, output.file.ext, tz)
}
\arguments{
\item{file.dir}{File directory or filename of single spectra for processing}

\item{out.dir}{Output directory for meta-data information file}

\item{spec.file.ext}{[Optional] Input spectra file extension. E.g. .asd (ASD) or .sed (Spectral Evolution).
Default for ASD instruments is .asd.  Default for Spectral Evolution instruments is .sed}

\item{output.file.ext}{[Optional] Output file extension of meta-data information file. Default .csv}

\item{tz}{[Optional] Set the timezone of the spectra file collection.  Used to covert spectra collection 
time to UTC.  If unused it is assumed that the correct timezone is the current system timezone.}
}
\value{
output Returns output dataframe of ASD metadata information
}
\description{
Extract metadata from raw binary ASD files. Called from extract.metadata
}
\author{
Shawn P. Serbin
}
