\name{extract.metadata.se}
\alias{extract.metadata.se}
\title{Extract metadata from Spectral Evolution files.  Called from extract.metadata}
\usage{
  extract.metadata.se(file.dir, out.dir, spec.file.ext,
    output.file.ext, tz)
}
\value{
  output Returns output dataframe of SE metadata
  information
}
\description{
  Extract metadata from Spectral Evolution files.  Called
  from extract.metadata
}
\author{
  Shawn P. Serbin
}

