% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.utils.R
\name{extract.metadata.svc}
\alias{extract.metadata.svc}
\title{Extract metadata from Spectra Vista files.  Called from extract.metadata}
\usage{
extract.metadata.svc(file.dir, out.dir, spec.file.ext, output.file.ext, tz)
}
\value{
output Returns output dataframe of SVC metadata information
}
\description{
Extract metadata from Spectra Vista files.  Called from extract.metadata
}
\author{
Shawn P. Serbin
}
