% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sig.R
\name{read.sig}
\alias{read.sig}
\title{read .sig files from SVC}
\usage{
read.sig(path, default = TRUE)
}
\arguments{
\item{path}{A character string specifying the path to the directory or a single sig file.}

\item{default}{A logical value indicating whether to include all columns from the `.sig` files
(`TRUE`) or only selected columns (`FALSE`). Default is `TRUE`.}
}
\value{
A data data frame with the following columns :
\describe{
\item{Wavelengths(nm)}{Numeric vector representing the wavelengths in nanometers.}
\item{file}{gives information on the file name}
\item{Reflectance(\%)}{Numeric vector of reflectance percentages.}
\item{Target Values (optional)}{Provides the targer values}
\item{Reference Values (optional)}{Provides information on the reference values}
}
}
\description{
Function to read .sig files and extract spectral data
}
\details{
This function will read ".sig" files and extract the speczral data. The resulting data frame will
have the option for all data as well as selected data , which can be obtained by providing the
boolean parameter.


The function can read a single file or multiple files in a directory.
The file is then read line by line and the spectral data is extracted from the line after "data=".
The extracted data is then converted into a data frame with appropriate names.
}
\examples{
\dontrun{
path <- "/path/to/your/file.sig"
data <- read.sig(path, default = FALSE)
print(head(data))
}

}
\author{
Methun George, Steffen Neumann
}
