% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.average.R
\name{spec.avg}
\alias{spec.avg}
\title{Average associated spectra files into a single spectrum. Works on a single spectra or a 
directory.}
\usage{
spec.avg(
  file.dir = NULL,
  out.dir = NULL,
  spec.type = "Reflectance",
  instrument = NULL,
  spec.file.ext = NULL,
  start.wave = NULL,
  end.wave = NULL,
  step.size = NULL,
  bias.threshold = NULL,
  outlier.cutoff = 2,
  suffix.length = NULL,
  output.file.ext = NULL,
  settings.file = NULL
)
}
\arguments{
\item{file.dir}{File directory or filename of single spectra for processing}

\item{out.dir}{Output directory for metadata information file}

\item{spec.type}{Option to set what type of spectra to process. Options: Reflectance,
Transmittance.  Can be set with abbreviations: e.g. "Refl" or "Tran". Default is "Reflectance"}

\item{instrument}{What instrument was used to collect spectra.  Current options: ASD, SE, SVC}

\item{spec.file.ext}{[Optional] Input spectra file extension. E.g. .asd (ASD), .sed (Spectral Evolution),
or .sig (Spectra Vista). In not input extension is assumed based on instrument type.}

\item{start.wave}{Starting wavelength of spectra files. 
Not needed if specified in XML settings file.}

\item{end.wave}{Ending wavelength of spectra files. Not needed if 
specified in XML settings file.}

\item{step.size}{Resolution of spectra files. E.g. 1 for 1nm, 5 for 5nm. 
Not needed if specified in XML settings file.  **Phasing this option out through the use of
spec file metadata and determining wavelength numbers from spec files}
}
\description{
Average associated spectra files into a single spectrum. Works on a single spectra or a 
directory.
}
